refactor: update promo code types and imports across billing and pricing modules for improved consistency and clarity
